Split dark mode into its own function
