+++ src/settings.ts
@@ -2,4 +2,5 @@ export interface Settings {
   autosaveSetting: boolean;
   originalGraphHasSet: number;
   graphHasSet: number;
+  skywardDarkTheme: boolean;
 }
